+++ R/initialise.R
@@ -11,11 +11,12 @@
 #'   system.file("examples", "dettl_config.yml", package = "dettl"),
 #'   add_log_table = FALSE
 #' )
-#' dettl::dettl_db_create_log_table(file.path(path, "person_information"), "test")
+#' dettl::dettl_db_create_log_table(file.path(path, "person_information"),
+#' "test")
 #'
 dettl_db_create_log_table <- function(path, db_name) {
   ## Get the connection info
-  path <- normalizePath(path, winslash = '/', mustWork = TRUE)
+  path <- normalizePath(path, winslash = "/", mustWork = TRUE)
   con <- db_connect(db_name, path)
   on.exit(DBI::dbDisconnect(con))
   ## Create the table