+++ R/load_runner.R
@@ -5,7 +5,8 @@
 #'
 #' @param load The load function for making the DB changes.
 #' @param con Connection to the database.
-#' @param extracted_data Extracted data if needed for testing, otherwise can be NULL.
+#' @param extracted_data Extracted data if needed for testing, otherwise can be
+#' NULL.
 #' @param transformed_data List of data frames representing the data to be
 #' loaded to the DB.
 #' @param test_queries Function containing queries for running on the DB before
@@ -22,17 +23,28 @@
 #' @keywords internal
 #'
 run_load <- function(con, load, extracted_data, transformed_data, test_queries,
-                     pre_load, post_load, path, test_file, dry_run, log_table,
-                     comment) {
+                     pre_load, post_load, path, test_file, transaction, dry_run,
+                     log_table, comment) {
+  if (is.null(transformed_data)) {
+    stop("Cannot run tests as no data has been transformed.")
+  }
   log_data <- build_log_data(path, comment)
   verify_log_table(con, log_table, log_data)
   verify_first_run(con, log_table, log_data)
-  DBI::dbBegin(con)
+  use_transaction <- transaction || dry_run
+  if (use_transaction) {
+    DBI::dbBegin(con)
+  }
   withCallingHandlers(
     do_load(con, load, extracted_data, transformed_data, path, test_file,
-            test_queries, pre_load, post_load, log_table, log_data, dry_run),
+            test_queries, pre_load, post_load, log_table, log_data, transaction,
+            dry_run),
     error = function(e) {
-      DBI::dbRollback(con)
+      if (use_transaction) {
+        DBI::dbRollback(con)
+      } else {
+        message("ATTENTION: even though your load has failed, because you did not use a transaction, the database may have been modified")
+      }
       stop(e)
     }
   )
@@ -41,25 +53,29 @@ run_load <- function(con, load, extracted_data, transformed_data, test_queries,
 
 do_load <- function(con, load, extracted_data, transformed_data, path,
                     test_file, test_queries, pre_load, post_load, log_table,
-                    log_data, dry_run) {
-  message("Running load:")
+                    log_data, transaction, dry_run) {
+  message(
+    sprintf("Running load %s:",
+            transaction %?% "in a transaction" %:% "not in a transaction"))
   message("\t- Running test queries before making any changes")
-  before <- test_queries(con)
-  if (!is.null(pre_load)) {
-    message("\t- Running pre-load")
-    pre_load(transformed_data, con)
-  }
-  message("\t- Running load step")
-  load(transformed_data, con)
-  if (!is.null(post_load)) {
-    message("\t- Running post-load")
-    post_load(transformed_data, con)
-  }
-  message("\t- Running test queries after making changes")
-  after <- test_queries(con)
-  test_path <- file.path(path, test_file)
-  message(sprintf("\t- Running load tests %s", test_path))
-  test_results <- run_load_tests(test_path, before, after, extracted_data, transformed_data, con)
+  withr::with_dir(path, {
+    before <- test_queries(con)
+    if (!is.null(pre_load)) {
+      message("\t- Running pre-load")
+      pre_load(transformed_data, con)
+    }
+    message("\t- Running load step")
+    load(transformed_data, con)
+    if (!is.null(post_load)) {
+      message("\t- Running post-load")
+      post_load(transformed_data, con)
+    }
+    message("\t- Running test queries after making changes")
+    after <- test_queries(con)
+    message(sprintf("\t- Running load tests %s", test_file))
+    test_results <- run_load_tests(test_file, before, after, extracted_data,
+                                   transformed_data, con)
+  })
   if (all_passed(test_results)) {
     if (dry_run) {
       DBI::dbRollback(con)
@@ -67,7 +83,9 @@ do_load <- function(con, load, extracted_data, transformed_data, path,
     } else {
       message("All tests passed, commiting changes to database.")
       write_log(con, log_table, log_data)
-      DBI::dbCommit(con)
+      if (transaction) {
+        DBI::dbCommit(con)
+      }
     }
   } else {
     stop("Failed to load data - not all tests passed.")