+++ tests/testthat/example_relative_path/R/test_transform.R
@@ -1,0 +1,5 @@
+context("test-transform")
+
+testthat::test_that("transformed data contains 2 rows", {
+  expect_equal(nrow(transformed_data$people), 2)
+})