+++ tests/testthat/test-utils-assert.R
@@ -5,7 +5,9 @@ test_that("can check for a function within an environment", {
   expect_error(assert_func_exists("test", env),
                "Can't find function test within environment.")
 
-  env$test <- function(){ "test func" }
+  env$test <- function() {
+    "test func"
+  }
   assert_func_exists("test", env)
   expect_equal(env$test(), "test func")
 })